cloudbuild/foundation/requirements.txt
google-auth==2.6.2
requests==2.27.1